mtest: refactor mtest_thread.c
Split each thread package into its own file.
